switch tech stack to spring and vue
