--- app.vue
@@ -1,7 +1,0 @@
-<script lang="ts" setup>
-import "./assets/tailwind.scss";
-</script>
-
-<template>
-  <NuxtPage />
-</template>
