--- assets/_custom.scss
@@ -1,5 +1,0 @@
-// custom styles while using tailwind theme
-
-// using tailwind theme variables inside scss with the theme() function
-// https://tailwindcss.com/docs/using-with-preprocessors#variables
-// https://tailwindcss.com/docs/functions-and-directives#theme
