--- assets/_mixins.scss
@@ -1,19 +1,0 @@
-// mixins
-
-// using tailwind theme variables inside scss with the theme() function
-// https://tailwindcss.com/docs/using-with-preprocessors#variables
-// https://tailwindcss.com/docs/functions-and-directives#theme
-
-$breakpoints: (
-  sm: theme("screens.sm"),
-  md: theme("screens.md"),
-  lg: theme("screens.lg"),
-  xl: theme("screens.xl"),
-);
-
-@mixin breakpoint($size) {
-  @media screen and (min-width: map-get($breakpoints, $size)) {
-    // insert content inside
-    @content;
-  }
-} ;
