+++ backend/src/main/java/dev/cuvar/panal/PanalApplication.java
@@ -1,0 +1,13 @@
+package dev.cuvar.panal;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PanalApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PanalApplication.class, args);
+	}
+
+}
