+++ backend/src/test/java/dev/cuvar/panal/PanalApplicationTests.java
@@ -1,0 +1,13 @@
+package dev.cuvar.panal;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PanalApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
