--- components/SiteStructure/Footer.vue
@@ -1,18 +1,0 @@
-<template>
-  <div class="w-full flex justify-center py-10">
-    <p>{{ title }}</p>
-  </div>
-</template>
-
-<script lang="ts">
-import { defineComponent } from "vue";
-
-const Footer = defineComponent({
-  data() {
-    return {
-      title: "panal" as string,
-    };
-  },
-});
-export default Footer;
-</script>
