+++ frontend/src/App.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+import { RouterView } from "vue-router";
+</script>
+
+<template>
+  <RouterView />
+</template>
