+++ frontend/src/assets/base.css
@@ -1,0 +1,4 @@
+/* color palette from <https://github.com/vuejs/theme> */
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
