+++ frontend/src/components/PageContent/WidgetContainer.vue
