+++ frontend/src/router/index.ts
@@ -1,0 +1,15 @@
+import { createRouter, createWebHistory } from "vue-router";
+import HomeView from "../views/index.vue";
+
+const router = createRouter({
+  history: createWebHistory(import.meta.env.BASE_URL),
+  routes: [
+    {
+      path: "/",
+      name: "home",
+      component: HomeView,
+    },
+  ],
+});
+
+export default router;
