+++ frontend/utils/mockdata/WidgetData.vue
