+++ frontend/utils/types/WidgetData.vue
