--- nuxt.config.ts
@@ -1,15 +1,0 @@
-import { defineNuxtConfig } from "nuxt3";
-
-export default defineNuxtConfig({
-  srcDir: "./",
-  build: {
-    postcss: {
-      postcssOptions: {
-        plugins: {
-          tailwindcss: {},
-          autoprefixer: {},
-        },
-      },
-    },
-  },
-});
