--- package.json
@@ -1,21 +1,0 @@
-{
-  "version": "1.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "nuxi dev",
-    "tw": "nuxi dev -o & tailwind-config-viewer -o -p 4000",
-    "build": "nuxi build",
-    "start": "node .output/server/index.mjs"
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.4.2",
-    "nuxt3": "^3.0.0-27251455.110b2b2",
-    "postcss": "^8.4.6",
-    "sass": "^1.43.2",
-    "tailwind-config-viewer": "^1.6.2",
-    "tailwindcss": "^3.0.18"
-  },
-  "engines": {
-    "node": ">=14.18.1"
-  }
-}
